+++ src/WeeklyQuiz/java_interface/DeliveryChargeCalculator.java
@@ -3,5 +3,5 @@ package WeeklyQuiz.java_interface;
 import java.math.BigDecimal;
 
 public interface DeliveryChargeCalculator {
-    BigDecimal getDeliveryCharge(int weight, BigDecimal price);
+    BigDecimal getDeliveryCharge(double weight, BigDecimal price);
 }
