+++ src/WeeklyQuiz/java_interface/Product.java
@@ -5,9 +5,9 @@ import java.math.BigDecimal;
 public abstract class Product implements Promotion{
     private String name;
     private BigDecimal price;
-    private int weight;
+    private double weight;
 
-    public Product(String name, BigDecimal price, int weight) {
+    public Product(String name, BigDecimal price, double weight) {
         this.name = name;
         this.price = price;
         this.weight = weight;
@@ -18,7 +18,7 @@ public abstract class Product implements Promotion{
     public BigDecimal getDiscountedPrice() {
         return this.price.subtract(getDiscountAmount());
     }
-    public int getWeight() {
+    public double getWeight() {
         return this.weight;
     }
 }
